--- lesson4/httptest/models.go
@@ -1,7 +1,0 @@
-package main
-
-type Employee struct {
-	Name   string  `json:"name" xml:"name"`
-	Age    int     `json:"age" xml:"age"`
-	Salary float32 `json:"salary" xml:"salary"`
-}
