--- lesson4/main.go
@@ -1,105 +1,0 @@
-package main
-
-import (
-	"encoding/json"
-	"encoding/xml"
-	"fmt"
-	"io"
-	"io/ioutil"
-	"log"
-	"net/http"
-	"time"
-)
-
-type Handler struct{}
-
-type UploadHandler struct {
-	HostAddr  string
-	UploadDir string
-}
-
-func (u *UploadHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	file, header, err := r.FormFile("file")
-	if err != nil {
-		http.Error(w, "Unable to read file", http.StatusBadRequest)
-		return
-	}
-	defer file.Close()
-
-	data, err := io.ReadAll(file)
-	if err != nil {
-		http.Error(w, "Unable to read file", http.StatusBadRequest)
-	}
-
-	filePath := u.UploadDir + "/" + header.Filename
-
-	err = ioutil.WriteFile(filePath, data, 0777)
-	if err != nil {
-		log.Println(err)
-		http.Error(w, "Unable to save file", http.StatusInternalServerError)
-		return
-	}
-
-	fileLink := u.HostAddr + "/" + header.Filename
-	fmt.Fprintln(w, fileLink)
-}
-
-func (h *Handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	switch r.Method {
-	case http.MethodGet:
-		name := r.FormValue("name")
-		fmt.Fprintf(w, "Parsed query-param with key \"name\": %s", name)
-	case http.MethodPost:
-		var employee Employee
-
-		contentType := r.Header.Get("Content-Type")
-
-		switch contentType {
-		case "application/json":
-			err := json.NewDecoder(r.Body).Decode(&employee)
-			if err != nil {
-				http.Error(w, "Unable to unmarshal JSON", http.StatusBadRequest)
-				return
-			}
-		case "application/xml":
-			err := xml.NewDecoder(r.Body).Decode(&employee)
-			if err != nil {
-				http.Error(w, "Unable to unmarshal XML", http.StatusBadRequest)
-				return
-			}
-
-		default:
-			http.Error(w, "Unknown content type", http.StatusBadRequest)
-			return
-		}
-
-		fmt.Fprintf(w, "Got a new employee!\nName: %s\nAge: %d y.o.\nSalary %0.2f\n",
-			employee.Name,
-			employee.Age,
-			employee.Salary,
-		)
-	}
-}
-
-func main() {
-	uploadHandler := &UploadHandler{
-		UploadDir: "./upload",
-		HostAddr:  "http://localhost:3000",
-	}
-	dirToServe := http.Dir(uploadHandler.UploadDir)
-
-	http.Handle("/", http.FileServer(dirToServe))
-	http.Handle("/upload", uploadHandler)
-
-	srv := &http.Server{
-		Addr:         "localhost:3000",
-		ReadTimeout:  10 * time.Second,
-		WriteTimeout: 10 * time.Second,
-	}
-
-	err := srv.ListenAndServe()
-	if err != nil {
-		log.Println(err)
-		return
-	}
-}
